Adding the first model of binary trie.
I'm using a modification of the original trie presented on the slides because the number of characters in the morse code can vary. The search is not finished yet. The code for search is the one that I was testing when I was trying to implement the trie presented on the slides. This code reads the morse.txt file and create a trie for it. The trie is printed after that, using the preorder tree traversal.